Some small modifications.
UnVals to UnDVs to be consistent with published papers.

modified the function names list so that in later versions of pin (2.9 onwards) .plt and other functions should not appear at the output.

modified the dot file printing such that the values are printed in separte lines on the edges. this is useful in complex applications and allows dot to produce more readable output.

removed the _ starting restriction from the unkown function name filter implementation, especially because KLT feature tracking library has functions (of our interest) name starting with underscores.

+++ Channel.cpp
@@ -16,7 +16,7 @@
 
 using namespace std;
 
-Channel::Channel(string p, string c, unsigned long unma, unsigned long bytes, unsigned long vals)
+Channel::Channel(string p, string c, unsigned long long unma, unsigned long long bytes, unsigned long long vals)
 {
 	producer = p;
 	consumer = c;
@@ -25,7 +25,7 @@ Channel::Channel(string p, string c, unsigned long unma, unsigned long bytes, un
 	Values = vals;
 }
 
-void Channel::setChannel(string p, string c, unsigned long unma, unsigned long bytes, unsigned long vals)
+void Channel::setChannel(string p, string c, unsigned long long unma, unsigned long long bytes, unsigned long long vals)
 {
 	setProducer(p);
 	setConsumer(c);
